+++ tests/IlluminaSampleSheet/V1/NovaSeqHeaderSectionTest.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace MLL\Utils\Tests\IlluminaSampleSheet;
+namespace MLL\Utils\Tests\IlluminaSampleSheet\V1;
 
 use PHPUnit\Framework\TestCase;
