+++ tests/IlluminaSampleSheet/V1/NovaSeqSampleSheetTest.php
@@ -1,23 +1,22 @@
 <?php declare(strict_types=1);
 
-namespace MLL\Utils\Tests\IlluminaSampleSheet;
+namespace MLL\Utils\Tests\IlluminaSampleSheet\V1;
 
-use MLL\Utils\IlluminaSampleSheet\V1\NovaSeqHeaderSection;
-use MLL\Utils\IlluminaSampleSheet\V1\NovaSeqXpDataSection;
-use MLL\Utils\IlluminaSampleSheet\V1\NovaSeqXpSample;
+use MLL\Utils\IlluminaSampleSheet\V1\NovaSeqSample;
+use MLL\Utils\IlluminaSampleSheet\V1\NovaSeqSampleSheet;
 use MLL\Utils\IlluminaSampleSheet\V1\SettingsSection;
 use PHPUnit\Framework\TestCase;
 
-class NovaSeqXpSampleSheetTest extends TestCase
+class NovaSeqSampleSheetTest extends TestCase
 {
     public function testNovaSeqSampleSheetAddsSectionsOnConstruction(): void
     {
-        $headerSection = $this->createMock(NovaSeqHeaderSection::class);
+        $headerSection = $this->createMock(\MLL\Utils\IlluminaSampleSheet\V1\NovaSeqHeaderSection::class);
         $readsSection = $this->createMock(\MLL\Utils\IlluminaSampleSheet\V1\ReadsSection::class);
-        $settingsSection = $this->createMock(SettingsSection::class);
+        $settingsSection = $this->createMock(\MLL\Utils\IlluminaSampleSheet\V1\SettingsSection::class);
         $dataSection = $this->createMock(\MLL\Utils\IlluminaSampleSheet\V1\DataSection::class);
 
-        $novaSeqSampleSheet = new \MLL\Utils\IlluminaSampleSheet\V1\NovaSeqXpSampleSheet($headerSection, $readsSection, $settingsSection, $dataSection);
+        $novaSeqSampleSheet = new NovaSeqSampleSheet($headerSection, $readsSection, $settingsSection, $dataSection);
 
         self::assertContains($headerSection, $novaSeqSampleSheet->getSections());
         self::assertContains($readsSection, $novaSeqSampleSheet->getSections());
@@ -27,7 +26,7 @@ class NovaSeqXpSampleSheetTest extends TestCase
 
     public function testNovaSeqStandardSampleSheetToStringReturnsExpectedResult(): void
     {
-        $headerSection = new NovaSeqHeaderSection(
+        $headerSection = new \MLL\Utils\IlluminaSampleSheet\V1\NovaSeqHeaderSection(
             '4',
             'DonalDuck',
             'MyExperiment',
@@ -41,11 +40,11 @@ class NovaSeqXpSampleSheetTest extends TestCase
 
         $readsSection = new \MLL\Utils\IlluminaSampleSheet\V1\ReadsSection(101, 101);
 
-        $dataSection = new NovaSeqXpDataSection();
+        $dataSection = new \MLL\Utils\IlluminaSampleSheet\V1\NovaSeqDataSection();
 
-        $dataSection->addSample(new NovaSeqXpSample('2', '1', 'Sample-001-M001', 'RunXXXX-PLATE', '', 'UDP0090', 'TCAGGCTT', 'UDP0090', 'ATCATGCG', 'RunXXXX-PROJECT', 'description'));
-        $dataSection->addSample(new NovaSeqXpSample('1', '2', 'Sample-002-M002', 'RunXXXX-PLATE', '', 'UDP0091', 'CCTTGTAG', 'UDP0091', 'CCTTGGAA', 'RunXXXX-PROJECT', 'description'));
-        $dataSection->addSample(new NovaSeqXpSample('4', '3', 'Sample-003-M003', 'RunXXXX-PLATE', '', 'UDP0092', 'GAACATCG', 'UDP0092', 'TCGACAAG', 'RunXXXX-PROJECT', 'description'));
+        $dataSection->addSample(new NovaSeqSample('1', 'Sample-001-M001', 'RunXXXX-PLATE', '', 'UDP0090', 'TCAGGCTT', 'UDP0090', 'ATCATGCG', 'RunXXXX-PROJECT', 'description'));
+        $dataSection->addSample(new NovaSeqSample('2', 'Sample-002-M002', 'RunXXXX-PLATE', '', 'UDP0091', 'CCTTGTAG', 'UDP0091', 'CCTTGGAA', 'RunXXXX-PROJECT', 'description'));
+        $dataSection->addSample(new NovaSeqSample('3', 'Sample-003-M003', 'RunXXXX-PLATE', '', 'UDP0092', 'GAACATCG', 'UDP0092', 'TCGACAAG', 'RunXXXX-PROJECT', 'description'));
 
         $settings = new SettingsSection('AGATCGGAAGAGCACACGTCTGAACTCCAGTCA', 'AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGT');
         $miSeqSampleSheet = new \MLL\Utils\IlluminaSampleSheet\V1\NovaSeqSampleSheet($headerSection, $readsSection, $settings, $dataSection);
@@ -67,10 +66,10 @@ Chemistry,MyChemistry
 Adapter,AGATCGGAAGAGCACACGTCTGAACTCCAGTCA
 AdapterRead2,AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGT
 [Data]
-Lane,Sample_ID,Sample_Name,Sample_Plate,Sample_Well,I7_Index_ID,Index,I5_Index_ID,Index2,Sample_Project,Description
-2,1,Sample-001-M001,RunXXXX-PLATE,,UDP0090,TCAGGCTT,UDP0090,ATCATGCG,RunXXXX-PROJECT,description
-1,2,Sample-002-M002,RunXXXX-PLATE,,UDP0091,CCTTGTAG,UDP0091,CCTTGGAA,RunXXXX-PROJECT,description
-4,3,Sample-003-M003,RunXXXX-PLATE,,UDP0092,GAACATCG,UDP0092,TCGACAAG,RunXXXX-PROJECT,description
+Sample_ID,Sample_Name,Sample_Plate,Sample_Well,I7_Index_ID,Index,I5_Index_ID,Index2,Sample_Project,Description
+1,Sample-001-M001,RunXXXX-PLATE,,UDP0090,TCAGGCTT,UDP0090,ATCATGCG,RunXXXX-PROJECT,description
+2,Sample-002-M002,RunXXXX-PLATE,,UDP0091,CCTTGTAG,UDP0091,CCTTGGAA,RunXXXX-PROJECT,description
+3,Sample-003-M003,RunXXXX-PLATE,,UDP0092,GAACATCG,UDP0092,TCGACAAG,RunXXXX-PROJECT,description
 ';
         self::assertSame($expected, $miSeqSampleSheet->toString());
     }
